+++ backend/bots/examples/sns1_airdrop/impl/src/updates/handle_direct_message.rs
@@ -22,6 +22,7 @@ fn handle_message(args: Args, state: &mut RuntimeState) -> Response {
             thread_root_message_id: None,
             content: MessageContentInitial::Text(TextContent { text }),
             message_id: None,
+            block_level_markdown: None,
         }],
     })
 }