+++ backend/canisters/community/api/src/updates/set_video_call_presence.rs
@@ -1,0 +1,22 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::{ChannelId, MessageId, VideoCallPresence};
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub channel_id: ChannelId,
+    pub message_id: MessageId,
+    pub presence: VideoCallPresence,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+    MessageNotFound,
+    AlreadyEnded,
+    CommunityFrozen,
+    UserNotInCommunity,
+    UserSuspended,
+    ChannelNotFound,
+    UserNotInChannel,
+}