+++ backend/canisters/user_index/impl/src/updates/c2c_report_message.rs
@@ -192,6 +192,7 @@ https://github.com/open-chat-labs/open-chat/commit/e93865ea29b5bab8a9f0b01052938
             thread_summary: None,
             edited: false,
             forwarded: false,
+            block_level_markdown: false,
         };
 
         let report = construct_html_report(Chat::Group(chat_id), None, &message, true);
@@ -227,6 +228,7 @@ https://github.com/open-chat-labs/open-chat/commit/e93865ea29b5bab8a9f0b01052938
             thread_summary: None,
             edited: false,
             forwarded: false,
+            block_level_markdown: false,
         };
 
         let report = construct_html_report(chat, None, &message, true);
@@ -266,6 +268,7 @@ https://github.com/open-chat-labs/open-chat/commit/e93865ea29b5bab8a9f0b01052938
             thread_summary: None,
             edited: false,
             forwarded: false,
+            block_level_markdown: false,
         };
 
         let report = construct_html_report(chat, None, &message, true);