+++ backend/integration_tests/src/client/group.rs
@@ -69,6 +69,7 @@ pub mod happy_path {
                 replies_to: None,
                 mentioned: Vec::new(),
                 forwarding: false,
+                block_level_markdown: false,
                 rules_accepted: None,
                 message_filter_failed: None,
                 correlation_id: 0,