+++ frontend/openchat-agent/src/services/user/user.client.ts
@@ -539,10 +539,16 @@ export class UserClient extends CandidService {
         event: EventWrapper<Message>,
         messageFilterFailed: bigint | undefined,
         threadRootMessageIndex: number | undefined,
-        pin: string | undefined,        
+        pin: string | undefined,
     ): Promise<[SendMessageResponse, Message]> {
         removeFailedMessage(this.db, this.chatId, event.event.messageId, threadRootMessageIndex);
-        return this.sendMessageToBackend(chatId, event, messageFilterFailed, threadRootMessageIndex, pin)
+        return this.sendMessageToBackend(
+            chatId,
+            event,
+            messageFilterFailed,
+            threadRootMessageIndex,
+            pin,
+        )
             .then(
                 setCachedMessageFromSendResponse(
                     this.db,
@@ -581,6 +587,7 @@ export class UserClient extends CandidService {
                 ),
                 forwarding: event.event.forwarded,
                 thread_root_message_index: apiOptional(identity, threadRootMessageIndex),
+                block_level_markdown: true,
                 message_filter_failed: apiOptional(identity, messageFilterFailed),
                 pin: apiOptional(identity, pin),
                 correlation_id: generateUint64(),
@@ -644,6 +651,7 @@ export class UserClient extends CandidService {
                 (replyContext) => apiReplyContextArgs(groupId, replyContext),
                 event.event.repliesTo,
             ),
+            block_level_markdown: true,
             message_filter_failed: apiOptional(identity, messageFilterFailed),
             pin: apiOptional(identity, pin),
             correlation_id: generateUint64(),
@@ -727,6 +735,7 @@ export class UserClient extends CandidService {
                 (replyContext) => apiReplyContextArgs(id, replyContext),
                 event.event.repliesTo,
             ),
+            block_level_markdown: true,
             community_rules_accepted: apiOptional(identity, communityRulesAccepted),
             channel_rules_accepted: apiOptional(identity, channelRulesAccepted),
             message_filter_failed: apiOptional(identity, messageFilterFailed),
@@ -1167,7 +1176,7 @@ export class UserClient extends CandidService {
         chatId: DirectChatIdentifier,
         threadRootMessageIndex: number | undefined,
         messageId: bigint,
-        deleteMessage: boolean,        
+        deleteMessage: boolean,
     ): Promise<boolean> {
         return this.handleResponse(
             this.userService.report_message({
@@ -1261,7 +1270,12 @@ export class UserClient extends CandidService {
         );
     }
 
-    acceptP2PSwap(userId: string, threadRootMessageIndex: number | undefined, messageId: bigint, pin: string | undefined): Promise<AcceptP2PSwapResponse> {
+    acceptP2PSwap(
+        userId: string,
+        threadRootMessageIndex: number | undefined,
+        messageId: bigint,
+        pin: string | undefined,
+    ): Promise<AcceptP2PSwapResponse> {
         return this.handleResponse(
             this.userService.accept_p2p_swap({
                 user_id: Principal.fromText(userId),